pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gwsdsc"
version = "0.1.0"
description = "Desired State Configuration framework for Google Workspace / Cloud Identity tenants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Your Organisation"}]
keywords = ["google-workspace", "cloud-identity", "dsc", "configuration-as-code", "iac"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "google-cloud-storage>=2.12.0",
    "google-cloud-access-context-manager>=0.2.0",
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "deepdiff>=6.7",
    "jinja2>=3.1",
    "gitpython>=3.1.40",
    "pydantic>=2.5",
]

[project.optional-dependencies]
azure = [
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
]
gcp-secrets = [
    "google-cloud-secret-manager>=2.17.0",
]
all-secrets = [
    "gwsdsc[azure,gcp-secrets]",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.1.6",
    "mypy>=1.7",
    "pre-commit>=3.6",
    "gwsdsc[all-secrets]",
]

[project.scripts]
gwsdsc = "gwsdsc.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
